Inline Demucs again
